refactor(mycarlos): extract modal focus handling
